Move test_cmdline tests to pytest and reorg some stuff
